Create paginate task widget
