+++ plugins/item_tasks/web_client/stylesheets/paginateTasksWidget.styl
@@ -1,0 +1,12 @@
+.g-task-list-header
+  margin-bottom 10px
+
+  .g-task-pagination
+    ul
+      margin 0px
+
+.g-execute-task-link
+    cursor pointer
+    
+.g-execute-task-link:hover
+    background-color #f0f0f0
