--- plugins/item_tasks/web_client/stylesheets/taskList.styl
@@ -1,6 +1,0 @@
-.g-task-list-header
-  margin-bottom 10px
-
-  .g-task-pagination
-    ul
-      margin 0px
