+++ plugins/item_tasks/web_client/templates/paginateTasksWidget.pug
@@ -3,6 +3,6 @@
 
 .list-group.g-task-list-container
   each task in tasks
-    a.list-group-item.g-execute-task-link(href=`#item_task/${task.id}/run`)
+    .list-group-item.g-execute-task-link(data-task-id=task.id)
       h4.list-group-item-heading.g-execute-task-link-header= task.name()
       .list-group-item-text.g-execute-task-link-body= task.get('description')
